src/main/java/com/anymind/pos/config/serialization/PriceSerializer.java
package com.anymind.pos.config.serialization;

import com.fasterxml.jackson.core.JsonGenerator;
import com.fasterxml.jackson.databind.JsonSerializer;
import com.fasterxml.jackson.databind.SerializerProvider;

import java.io.IOException;
import java.math.BigDecimal;
import java.math.RoundingMode;

public class PriceSerializer extends JsonSerializer<BigDecimal> {

    @Override
    public void serialize(BigDecimal value, JsonGenerator gen, SerializerProvider serializers) throws IOException {
        if (value != null) {
            BigDecimal formattedValue = value.setScale(2, RoundingMode.HALF_UP);
            gen.writeNumber(formattedValue);
        } else {
            gen.writeNull();
        }
    }
}
